Added an enum with the different deviceTypes in use by the software and the corresponding variables on the concrete classes, this is being used in main.cpp to decide which kind of device is instantiated and call the corresponding method for each.
Added a pointer to the spatial_hashing into the haptic device class to handle
the collision detection and collision response.

Added a force scaler to downscale the forces produced by the haptic device.

Now passing a position to the 'getForceFeedback' function.

FIX: Should move the m_spaceHandler variable to LYHapticInterface.h as protected

Finalized the collision response function using the Leeper, Chan and Salisbury
technique described in 'Point Clouds as Implicit Surfaces for Constraint-Based
Haptic Rendering', using in particular the Wendland weighting function.

Added the LYHapticDevice class responsable for doing the haptic rendering and
interaction with the haptic device.