Changed the object used for collision detection to something that makes more sense, storing the surface normal, surface contact point and haptic position.
Using assimp to read polygonal models and rply to read point data.

Moved all the haptic_interface interactions to a bigger class that will handle
the haptic environment (ie, moving the haptic workspace, etc).

TODO:
The overlay renderer is completed, now I have to find a way to connect it with the
haptic_interface in order to render the force vector / haptic workspace /etc on
top of the point renderer.