LYCudaHelper: commented line trying to solve the critical section bug.
LYHapticDevice: moved the timing code to the parent class in order to register
the timing for all the haptic devices in a compatible way.

LYHapticInterface: received the inteface timing code.

LYKeyboardDevice: Notices the shell when it's being used as the haptic interface.

LYSpatialHash: Trying to solve the problem with the surface translation.

main: Added code to file register the timing measures for the haptic interface.