Making sure all the models have normals.
Changed the light position on the final pass 'shader.frag'.

Widened the blur pass windowWidth within 'blur_pass.frag'.

Added ModelInfo.h, to hold the configuration of the models that are being loaded
also changed a bit the model loading functions to account for the loading of the
parameters and refactored some code that was repeated in model loading purposes.
All of these changes happened in main.cpp.

Making sure the colliders are visible again by calling _drawCollder() in the
surface rendering method. LYScreenspaceRenderer.cpp